Magik commands now automatically reduce the size of the image until its an acceptable size

+++ javascript/commands/magik_commands/inflate.js
@@ -4,8 +4,11 @@ const request = require("request");
 const rp = require("request-promise");
 
 const utilitiesModule = require('../../utilities');
+const magikUtilities = require('../../magikUtilities');
 const ahm = require("../../achievementHandler");
 const config = require("../../../data/general_data/config.json");
+
+const maxFileSize = 2;
 
 
 
@@ -59,11 +62,9 @@ module.exports.run = async (bot, message, args) => {
 
 
 
-    let foundURL;
-
     utilitiesModule.getMostRecentImageURL(message).then(returnedURL => {
 
-        foundURL = returnedURL;
+        let foundURL = returnedURL;
 
         if (!foundURL) {
             return;
@@ -76,40 +77,58 @@ module.exports.run = async (bot, message, args) => {
 
             rp(options)
                 .then(function (response) {
+
+                    let filename = Date.now();
                     let fileSize = (response.headers['content-length'] / 1000000.0).toFixed(2);
 
-                    if (fileSize > 2) {
-                        message.channel.send(`I don't want to fuck with anything around the size of 2mb, ${utilitiesModule.getRandomNameInsult(message)}`);
-                        return;
+                    //If we need to do recursive filesize reduction, we need to write the file to disk first
+                    if (fileSize > maxFileSize && inflateAmount != 69) {
+                        message.channel.send(`This image is **~${fileSize}mb**, I gotta chop it down until it's lower than **${maxFileSize}mb** (might take a bit, be patient)`);
+                        gm(request(foundURL))
+                            .write(`./graphics/${filename}.png`, function (err) {
+                                if (err) console.error(err);
+
+                                magikUtilities.reduceImageFileSize(message, filename, 1, maxFileSize, () => {
+                                    gm(`./graphics/${filename}.png`)
+                                        .implode(-inflateAmount)
+                                        .write(`./graphics/${filename}.png`, function (err) {
+                                            if (err) console.error(err);
+
+                                            message.channel.send({ files: [`./graphics/${filename}.png`] })
+                                                .then(function(msg) {
+                                                    fs.unlink(`./graphics/${filename}.png`, function(err) { if (err) throw err; });
+                                                })
+                                                .catch(console.error);
+                                        });
+                                });
+                            });
                     }
+
+                    //Otherwise, just do magik on the image directly
                     else {
                         let msg = `alright hold on, inflating a ~${fileSize}mb image`;
                         if (appendSuggestion) msg += ` (for best results, keep inflation strength less than 2)`;
                         message.channel.send(msg);
-                        
-                        let filename = Date.now();
 
-                        gm(request(foundURL))
-                            .implode(-inflateAmount)
-                            .write(`./graphics/${filename}.png`, function (err) {
-                                if (err) console.error(err);
-                                if (inflateAmount == 69) {
-                                    message.channel.send({ files: [`./graphics/misc/gotcha.png`] })
-                                        .then(function(msg) {
-                                            fs.unlink(`./graphics/${filename}.png`, function(err) { if (err) throw err; });
-                                        })
-                                        .catch(console.error);
-                                    ahm.awardAchievement(message, ahm.achievement_list_enum.SECRET_PORYGON);
-                                }
-                                else {
+                        if (inflateAmount == 69) {
+                            message.channel.send({ files: [`./graphics/${filename}.png`] });
+                            ahm.awardAchievement(message, ahm.achievement_list_enum.SECRET_PORYGON);
+                        }
+                        else {
+                            gm(request(foundURL))
+                                .implode(-inflateAmount)
+                                .write(`./graphics/${filename}.png`, function (err) {
+                                    if (err) console.error(err);
+
                                     message.channel.send({ files: [`./graphics/${filename}.png`] })
                                         .then(function(msg) {
                                             fs.unlink(`./graphics/${filename}.png`, function(err) { if (err) throw err; });
                                         })
                                         .catch(console.error);
-                                }
-                            });
+                                });
+                        }
                     }
+
                 })
                 .catch(function (err) {
                     console.error(err);
